requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.1
redis==5.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.1
pytest==7.4.4
websockets==12.0
email-validator